Moves tests to proper package and adds one more test case
